replace vuu-data dependencies with local or remote (#1108)
