+++ vuu-ui/packages/vuu-data-ag-grid/package.json
@@ -9,8 +9,8 @@
     "type-defs": "node ../../scripts/build-type-defs.mjs"
   },
   "dependencies": {
-    "@finos/vuu-data": "0.0.26",
     "@finos/vuu-data-react": "0.0.26",
+    "@finos/vuu-data-remote": "0.0.26",
     "@finos/vuu-data-types": "0.0.26",
     "@finos/vuu-table-types": "0.0.26",
     "@finos/vuu-filter-types": "0.0.26",
