+++ vuu-ui/packages/vuu-data-react/package.json
@@ -16,7 +16,7 @@
     "@finos/vuu-protocol-types": "0.0.26"
   },
   "dependencies": {
-    "@finos/vuu-data": "0.0.26",
+    "@finos/vuu-data-remote": "0.0.26",
     "@finos/vuu-data-types": "0.0.26",
     "@finos/vuu-filter-parser": "0.0.26",
     "@finos/vuu-utils": "0.0.26"
