+++ vuu-ui/packages/vuu-data-remote/src/message-utils.ts
@@ -74,24 +74,6 @@ export interface VuuTableMetaWithTable extends VuuTableMeta {
   table: VuuTable;
 }
 
-export const getColumnByName = (
-  schema: TableSchema,
-  name?: string
-): SchemaColumn | undefined => {
-  if (name === undefined) {
-    return undefined;
-  } else {
-    const column = schema.columns.find((col) => col.name === name);
-    if (column) {
-      return column;
-    } else {
-      throw Error(
-        `getColumnByName no column '${name}' in schema for ${schema.table.table}`
-      );
-    }
-  }
-};
-
 export const createSchemaFromTableMetadata = ({
   columns,
   dataTypes,
