+++ vuu-ui/packages/vuu-datatable/package.json
@@ -9,7 +9,7 @@
     "type-defs": "node ../../scripts/build-type-defs.mjs"
   },
   "dependencies": {
-    "@finos/vuu-data": "0.0.26",
+    "@finos/vuu-data-local": "0.0.26",
     "@finos/vuu-table-types": "0.0.26",
     "@finos/vuu-filters": "0.0.26",
     "@finos/vuu-layout": "0.0.26",
