+++ vuu-ui/packages/vuu-filters/package.json
@@ -17,7 +17,6 @@
   },
   "dependencies": {
     "@finos/vuu-codemirror": "0.0.26",
-    "@finos/vuu-data": "0.0.26",
     "@finos/vuu-data-react": "0.0.26",
     "@finos/vuu-popups": "0.0.26",
     "@finos/vuu-ui-controls": "0.0.26",
