+++ vuu-ui/packages/vuu-shell/package.json
@@ -11,7 +11,7 @@
   },
   "dependencies": {
     "@salt-ds/core": "1.13.2",
-    "@finos/vuu-data": "0.0.26",
+    "@finos/vuu-data-remote": "0.0.26",
     "@finos/vuu-icons": "0.0.26",
     "@finos/vuu-filters": "0.0.26",
     "@finos/vuu-layout": "0.0.26",
