+++ vuu-ui/sample-apps/app-vuu-example/package.json
@@ -19,7 +19,8 @@
     "@fontsource/open-sans": "^4.5.13",
     "@salt-ds/core": "1.13.2",
     "@salt-ds/lab": "1.0.0-alpha.15",
-    "@finos/vuu-data": "0.0.26",
+    "@finos/vuu-data-local": "0.0.26",
+    "@finos/vuu-data-remote": "0.0.26",
     "@finos/vuu-table-types": "0.0.26",
     "@finos/vuu-data-react": "0.0.26",
     "@finos/vuu-layout": "0.0.26",
