+++ vuu-ui/sample-apps/feature-basket-trading/package.json
@@ -21,7 +21,7 @@
     "@finos/vuu-table-types": "0.0.26"
   },
   "dependencies": {
-    "@finos/vuu-data": "0.0.26",
+    "@finos/vuu-data-remote": "0.0.26",
     "@finos/vuu-data-react": "0.0.26",
     "@finos/vuu-filters": "0.0.26",
     "@finos/vuu-layout": "0.0.26",
