+++ vuu-ui/sample-apps/feature-template/package.json
@@ -16,7 +16,7 @@
   ],
   "devDependencies": {},
   "dependencies": {
-    "@finos/vuu-data": "0.0.26",
+    "@finos/vuu-data-remote": "0.0.26",
     "@finos/vuu-data-react": "0.0.26",
     "@finos/vuu-table-types": "0.0.26",
     "@finos/vuu-filters": "0.0.26",
